+++ addons/main/magazines/CfgMagAK545New.hpp
@@ -1,0 +1,74 @@
+class hlc_30Rnd_545x39_Mdim_ak: hlc_30Rnd_545x39_m_ak {
+    ammo = "FH_545x39_Tracer_Dim";
+    descriptionShort = "Caliber: 5.45x39mm 7N6M/7T3MN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-74, RPK-74";
+    displayName = "30Rnd 5.45x39 7N6M Bakelite Mag (IR-DIM Tracer Mixed)";
+    displayNameShort = "FMJ/IR-DIM";
+};
+class hlc_30Rnd_545x39_Mdim_ak_Black: hlc_30Rnd_545x39_m_ak_Black {
+    ammo = "FH_545x39_Tracer_Dim";
+    descriptionShort = "Caliber: 5.45x39mm 7N6M/7T3MN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-74, RPK-74";
+    displayName = "30Rnd 5.45x39 7N6M Mag (IR-DIM Tracer Mixed)";
+    displayNameShort = "FMJ/IR-DIM";
+};
+class hlc_30Rnd_545x39_Mdim_ak_Plum: hlc_30Rnd_545x39_m_ak_Plum {
+    ammo = "FH_545x39_Tracer_Dim";
+    descriptionShort = "Caliber: 5.45x39mm 7N6M/7T3MN<br/>Type: FMJ/IR-DIM<br/>Rounds: 30<br/>Used in: AK-74, RPK-74";
+    displayName = "30Rnd 5.45x39 7N6M Plum Mag (IR-DIM Tracer Mixed)";
+    displayNameShort = "FMJ/IR-DIM";
+};
+class hlc_45Rnd_545x39_Mdim_rpk: hlc_45Rnd_545x39_m_rpk {
+    ammo = "FH_545x39_Tracer_Dim";
+    descriptionShort = "Caliber: 5.45x39mm 7N6M/7T3MN<br/>Type: FMJ/IR-DIM<br/>Rounds: 45<br/>Used in: AK-74, RPK-74";
+    displayName = "45Rnd 5.45x39 7N6M Bakelite Mag (IR-DIM Tracer Mixed)";
+    displayNameShort = "FMJ/IR-DIM";
+};
+class hlc_45Rnd_545x39_Mdim_rpkm: hlc_45Rnd_545x39_m_rpkm {
+    ammo = "FH_545x39_Tracer_Dim";
+    descriptionShort = "Caliber: 5.45x39mm 7N6M/7T3MN<br/>Type: FMJ/IR-DIM<br/>Rounds: 45<br/>Used in: AK-74, RPK-74";
+    displayName = "45Rnd 5.45x39 7N6M Mag (IR-DIM Tracer Mixed)";
+    displayNameShort = "FMJ/IR-DIM";
+};
+class hlc_60Rnd_545x39_Mdim_rpk: hlc_60Rnd_545x39_m_rpk {
+    ammo = "FH_545x39_Tracer_Dim";
+    descriptionShort = "Caliber: 5.45x39mm 7N6M/7T3MN<br/>Type: FMJ/IR-DIM<br/>Rounds: 60<br/>Used in: AK-74, RPK-74";
+    displayName = "60Rnd 5.45x39 7N6M Mag (IR-DIM Tracer Mixed)";
+    displayNameShort = "FMJ/IR-DIM";
+};
+
+
+class hlc_30Rnd_545x39_Tdim_ak: hlc_30Rnd_545x39_Mdim_ak {
+    descriptionShort = "Caliber: 5.45x39mm 7T3MN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-74, RPK-74";
+    displayName = "30Rnd 5.45x39 7T3MN Bakelite Mag (IR-DIM Tracer)";
+    displayNameShort = "IR-DIM";
+    tracersEvery = 1;
+};
+class hlc_30Rnd_545x39_Tdim_ak_Black: hlc_30Rnd_545x39_Mdim_ak_Black {
+    descriptionShort = "Caliber: 5.45x39mm 7T3MN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-74, RPK-74";
+    displayName = "30Rnd 5.45x39 7T3MN Mag (IR-DIM Tracer)";
+    displayNameShort = "IR-DIM";
+    tracersEvery = 1;
+};
+class hlc_30Rnd_545x39_Tdim_ak_Plum: hlc_30Rnd_545x39_Mdim_ak_Plum {
+    descriptionShort = "Caliber: 5.45x39mm 7T3MN<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-74, RPK-74";
+    displayName = "30Rnd 5.45x39 7T3MN Plum Mag (IR-DIM Tracer)";
+    displayNameShort = "IR-DIM";
+    tracersEvery = 1;
+};
+class hlc_45Rnd_545x39_Tdim_rpk: hlc_45Rnd_545x39_Mdim_rpk {
+    descriptionShort = "Caliber: 5.45x39mm 7T3MN<br/>Type: IR-DIM<br/>Rounds: 45<br/>Used in: AK-74, RPK-74";
+    displayName = "45Rnd 5.45x39 7T3MN Bakelite Mag (IR-DIM Tracer)";
+    displayNameShort = "IR-DIM";
+    tracersEvery = 1;
+};
+class hlc_45Rnd_545x39_Tdim_rpkm: hlc_45Rnd_545x39_Mdim_rpkm {
+    descriptionShort = "Caliber: 5.45x39mm 7T3MN<br/>Type: IR-DIM<br/>Rounds: 45<br/>Used in: AK-74, RPK-74";
+    displayName = "45Rnd 5.45x39 7T3MN Mag (IR-DIM Tracer)";
+    displayNameShort = "IR-DIM";
+    tracersEvery = 1;
+};
+class hlc_60Rnd_545x39_Tdim_rpk: hlc_60Rnd_545x39_Mdim_rpk {
+    descriptionShort = "Caliber: 5.45x39mm 7T3MN<br/>Type: IR-DIM<br/>Rounds: 60<br/>Used in: AK-74, RPK-74";
+    displayName = "60Rnd 5.45x39 7T3MN Mag (IR-DIM Tracer)";
+    displayNameShort = "IR-DIM";
+    tracersEvery = 1;
+};