+++ addons/main/magazines/CfgMagAK556.hpp
@@ -1,0 +1,73 @@
+class hlc_30rnd_556x45_S;
+class hlc_30Rnd_556x45_EPR_AK: hlc_30rnd_556x45_EPR {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm M855A1<br/>Type: EPR<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 M855A1 Mag";
+    displayNameShort = "EPR";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_FMJ_CA.paa";
+};
+class hlc_30Rnd_556x45_MDim_AK: hlc_30rnd_556x45_MDim {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm M855A1/XM996<br/>Type: EPR/IR-DIM<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 M855A1 Mag (IR-DIM Tracer Mixed)";
+    displayNameShort = "EPR/IR-DIM";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_Mixed_CA.paa";
+};
+class hlc_30Rnd_556x45_M_AK: hlc_30rnd_556x45_M {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm M855A1/M856A1<br/>Type: EPR/Tracer<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 M855A1 Mag (Red Tracer Mixed)";
+    displayNameShort = "EPR/Tracer";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_Mixed_CA.paa";
+};
+class hlc_30Rnd_556x45_TDim_AK: hlc_30rnd_556x45_TDim {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm XM996<br/>Type: IR-DIM<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 XM996 Mag (IR-DIM Tracer)";
+    displayNameShort = "IR-DIM";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_Tracer_CA.paa";
+};
+class hlc_30Rnd_556x45_SPR_AK: hlc_30rnd_556x45_SPR {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm Mk262 Mod 1<br/>Type: SBLR<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 Mk262 Mod 1 Mag";
+    displayNameShort = "SBLR";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_FMJ_CA.paa";
+};
+class hlc_30Rnd_556x45_SOST_AK: hlc_30rnd_556x45_SOST {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm Mk318 Mod 0<br/>Type: SOST<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 Mk318 Mod 0 Mag";
+    displayNameShort = "SOST";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_FMJ_CA.paa";
+};
+class hlc_30Rnd_556x45_S_AK: hlc_30rnd_556x45_S {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm Subsonic<br/>Type: SD<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 Subsonic Mag";
+    displayNameShort = "SD";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_Subsonic_CA.paa";
+};
+class hlc_30Rnd_556x45_T_AK: hlc_30rnd_556x45_t {
+    author = "NIArms & johnb43";
+    descriptionShort = "Caliber: 5.56x45mm M856A1<br/>Type: Tracer<br/>Rounds: 30<br/>Used in: AK-101, AEK-972";
+    displayName = "30Rnd 5.56x45 M856A1 Mag (Red Tracer)";
+    displayNameShort = "Tracer";
+    model = "\hlc_core\mesh\magazines\30rnd_545x39_AK74.p3d";
+    modelSpecial = "\hlc_core\mesh\magazines\proxies\30rnd_545x39_AK74_6L23";
+    picture = "\hlc_core\tex\ui\ammo\M_AK_6L23_Tracer_CA.paa";
+};