+++ addons/main/weapons/CfgWeapAK.hpp
@@ -1,35 +1,48 @@
+class UGL_F;
 class hlc_ak_base: Rifle_Base_F {
+    descriptionShort = "Infantry Rifle<br/>Caliber:5.45x39mm";
+    class hlc_GP30: UGL_F {
+        displayname = "GP-34";
+    };
+    class hlc_GP25: hlc_GP30 {
+        displayname = "GP-25";
+    };
     class WeaponSlotsInfo: WeaponSlotsInfo {};
 };
 
 class hlc_rifle_ak12: hlc_ak_base {
-    displayName = "Izhmash AK-12";
+    displayName = "AK-12";
+    magazineWell[] += {"AK_545x39"};
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 77;
     };
 };
 class hlc_rifle_ak12GL: hlc_rifle_ak12 {
-    displayName = "Izhmash AK-12 (GL)";
+    displayName = "AK-12 (GP-34)";
     class WeaponSlotsInfo {
         mass = 110;
     };
 };
 class hlc_rifle_aku12: hlc_rifle_ak12 {
-    displayName = "Izhmash AK-12K";
+    descriptionShort = "SMG<br/>Caliber:5.45x39mm";
+    displayName = "AK-12K";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 71;
     };
 };
 class hlc_rifle_RPK12: hlc_rifle_ak12 {
-    displayName = "Izhmash RPK-16";
+    descriptionShort = "Light Support Weapon<br/>Caliber:5.45x39mm";
+    displayName = "RPK-16";
     class WeaponSlotsInfo: WeaponSlotsInfo {
+        mass = 98;
         class GripodSlot: WB_rifle_grips_slot {};
         class UnderBarrelSlot: WB_rifle_grips_slot {};
     };
 };
 
 class hlc_rifle_ak74: hlc_ak_base {
-    displayName = "Izhmash AK-74N";
+    displayName = "AK-74N";
+    magazineWell[] += {"AK_545x39"};
     rhs_1p29_type = "rhs_acc_1p29";
     rhs_pgo7v_type = "rhs_acc_pgo7v_ak";
     rhs_pgo7v2_type = "rhs_acc_pgo7v2_ak";
@@ -42,62 +55,64 @@ class hlc_rifle_ak74: hlc_ak_base {
     };
 };
 class hlc_rifle_ak74_dirty: hlc_rifle_ak74 {
-    displayName = "Izhmash AK-74N (Worn)";
+    displayName = "AK-74N (Worn)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 73;
     };
 };
 class hlc_rifle_ak74_dirty2: hlc_rifle_ak74 {
-    displayName = "Izhmash AK-74N (Worn)";
+    displayName = "AK-74N (Worn)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 73;
     };
 };
 class hlc_rifle_ak74_MTK: hlc_rifle_ak74 {
-    displayName = "Izhmash AK-74N (MTK)";
+    displayName = "AK-74N (MTK)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 77;
     };
 };
 class hlc_rifle_aks74: hlc_rifle_ak74 {
-    displayName = "Izhmash AKS-74N";
+    displayName = "AKS-74N";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 66;
     };
 };
 class hlc_rifle_aks74_GL: hlc_rifle_aks74 {
-    displayName = "Izhmash AKS-74N (GL)";
+    displayName = "AKS-74N (GP-34)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 99;
     };
 };
 class hlc_rifle_aks74_MTK: hlc_rifle_aks74 {
-    displayName = "Izhmash AKS-74N (MTK)";
+    displayName = "AKS-74N (MTK)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 70;
     };
 };
 class hlc_rifle_ak74m: hlc_rifle_ak74 {
-    displayName = "Izhmash AK-74M";
+    displayName = "AK-74M";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 79.5;
     };
 };
 class hlc_rifle_ak74m_gl: hlc_rifle_ak74m {
-    displayName = "Izhmash AK-74M (GL)";
+    displayName = "AK-74M (GP-34)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 112.5;
     };
 };
 class hlc_rifle_ak74m_MTK: hlc_rifle_ak74m {
-    displayName = "Izhmash AK-74M (MTK)";
+    displayName = "AK-74M (MTK)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 83.5;
     };
 };
 
 class hlc_rifle_aks74u: hlc_ak_base {
-    displayName = "Izhmash AKS-74UN";
+    descriptionShort = "SMG<br/>Caliber:5.45x39mm";
+    displayName = "AKS-74UN";
+    magazineWell[] += {"AK_545x39"};
     rhs_1p29_type = "rhs_acc_1p29";
     rhs_pgo7v_type = "rhs_acc_pgo7v_ak";
     rhs_pgo7v2_type = "rhs_acc_pgo7v2_ak";
@@ -110,14 +125,15 @@ class hlc_rifle_aks74u: hlc_ak_base {
     };
 };
 class hlc_rifle_aks74u_MTK: hlc_rifle_aks74u {
-    displayName = "Izhmash AKS-74UN (MTK)";
+    displayName = "AKS-74UN (MTK)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 64.6;
     };
 };
 
 class hlc_rifle_ak47: hlc_ak_base {
-    displayName = "Izhmash AK-47N";
+    displayName = "AK-47N";
+    magazineWell[] += {"AK_762x39"};
     rhs_1p29_type = "rhs_acc_1p29";
     rhs_pgo7v_type = "rhs_acc_pgo7v_ak";
     rhs_pgo7v2_type = "rhs_acc_pgo7v2_ak";
@@ -130,25 +146,27 @@ class hlc_rifle_ak47: hlc_ak_base {
     };
 };
 class hlc_rifle_akm: hlc_rifle_ak47 {
-    displayName = "Izhmash AKMN";
+    displayName = "AKMN";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 70.1;
     };
 };
 class hlc_rifle_akm_MTK: hlc_rifle_akm {
-    displayName = "Izhmash AKMN (MTK)";
+    displayName = "AKMN (MTK)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 74.1;
     };
 };
 class hlc_rifle_akmgl: hlc_rifle_akm {
-    displayName = "Izhmash AKMN (GL)";
+    displayName = "AKMN (GP-25)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 103.1;
     };
 };
 
 class hlc_rifle_slr107u: hlc_ak_base {
+    displayName = "SLR-107UR";
+    magazineWell[] += {"AK_762x39"};
     rhs_1p29_type = "rhs_acc_1p29";
     rhs_pgo7v_type = "rhs_acc_pgo7v_ak";
     rhs_pgo7v2_type = "rhs_acc_pgo7v2_ak";
@@ -161,35 +179,44 @@ class hlc_rifle_slr107u: hlc_ak_base {
     };
 };
 class hlc_rifle_slr107u_MTK: hlc_rifle_slr107u {
+    displayName = "SLR-107UR (MTK)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 65;
     };
 };
 
 class hlc_rifle_aek971: hlc_rifle_ak74 {
+    displayName = "AEK-971S";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 73;
     };
 };
 class hlc_rifle_aek971_mtk: hlc_rifle_aek971 {
+    displayName = "AEK-971S (MTK)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 77;
     };
 };
 class hlc_rifle_aek971worn: hlc_rifle_aek971 {
+    displayName = "AEK-971S (Worn)";
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 73;
     };
 };
 
 class hlc_rifle_RK62: hlc_ak_base {
+    displayName = "RK 62";
+    magazines[] = {"hlc_30Rnd_762x39_b_ak_Valmet"};
+    magazineWell[] += {"AK_762x39"};
+    delete nia_magSwitch;
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 77;
     };
 };
 
 class hlc_rifle_rpk: hlc_ak_base {
-    displayName = "Molot RPKN";
+    displayName = "RPKN";
+    magazineWell[] += {"AK_762x39"};
     rhs_1p29_type = "rhs_acc_1p29";
     rhs_pgo7v_type = "rhs_acc_pgo7v_ak";
     rhs_pgo7v2_type = "rhs_acc_pgo7v2_ak";
@@ -197,13 +224,14 @@ class hlc_rifle_rpk: hlc_ak_base {
     rhs_pkas_type = "rhs_acc_pkas";
     rhs_pso1m2_type = "rhs_acc_pso1m2";
     rhs_pso1m21_type = "rhs_acc_pso1m21_svd";
-    modes[] = {"FullAuto","Single","AI_long","AI_close","AI_short","AI_medium","AI_far","AI_toofar","AI_far_optic1","AI_toofar_optic1","AI_far_optic2","AI_toofar_optic2"};
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 107;
     };
 };
 class hlc_rifle_rpk74n: hlc_ak_base {
-    displayName = "Molot RPK-74N";
+    descriptionShort = "Light Support Weapon<br/>Caliber:5.45x39mm";
+    displayName = "RPK-74N";
+    magazineWell[] += {"AK_545x39"};
     rhs_1p29_type = "rhs_acc_1p29";
     rhs_pgo7v_type = "rhs_acc_pgo7v_ak";
     rhs_pgo7v2_type = "rhs_acc_pgo7v2_ak";
@@ -211,7 +239,6 @@ class hlc_rifle_rpk74n: hlc_ak_base {
     rhs_pkas_type = "rhs_acc_pkas";
     rhs_pso1m2_type = "rhs_acc_pso1m2";
     rhs_pso1m21_type = "rhs_acc_pso1m21_svd";
-    modes[] = {"FullAuto","Single","AI_long","AI_close","AI_short","AI_medium","AI_far","AI_toofar","AI_far_optic1","AI_toofar_optic1","AI_far_optic2","AI_toofar_optic2"};
     class WeaponSlotsInfo: WeaponSlotsInfo {
         mass = 101;
     };