+++ addons/main/weapons/CfgWeapMG.hpp
@@ -8,9 +8,14 @@ class HLC_wp_M134Painless: Rifle_Base_F {
     };
 };
 
-class hlc_M60e4_base: Rifle_Base_F {};
+class hlc_M60e4_base: Rifle_Base_F {
+    class WeaponSlotsInfo: WeaponSlotsInfo {};
+};
 class hlc_lmg_m60: hlc_M60e4_base {
     delete nia_magSwitch;
+    class WeaponSlotsInfo: WeaponSlotsInfo {
+        mass = 231;
+    };
 };
 class hlc_lmg_M60E4: hlc_M60e4_base {
     delete nia_magSwitch;